Add updated file name after first review
