+++ src/App.js
@@ -1,9 +1,17 @@
 import React from 'react';
-// import Calculator from './components/Calculator';
-import CalculatorFun from './components/CalculatorFun';
+import Calculator from './components/Calculator';
 
-const App = () => (
-  <CalculatorFun />
-);
+class App extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+  }
+
+  render() {
+    return (
+      <Calculator />
+    );
+  }
+}
 
 export default App;
