--- src/components/styles/CalculatorFun.css
@@ -1,63 +1,0 @@
-
-@import url('https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,900;1,700;1,900&display=swap');
-
-* {
-  font-family: 'Roboto', sans-serif;
-  font-size: 14px;
-  font-weight: 400;
-}
-
-body {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.container {
-  margin: auto;
-  width: 320px;
-  text-align: center;
-  border: 1px solid #666;
-  background-color: #666;
-  border-radius: 1px;
-}
-
-.keypad {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  grid-auto-rows: minmax(60px, auto);
-  border: none;
-}
-
-.screen-window {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  height: 60px;
-  width: 100%;
-  text-align: right;
-  background: rgb(209, 201, 201);
-  font-weight: 500;
-  font-size: 18px;
-  border: none;
-}
-
-.keypad button {
-  border: none;
-  border-bottom: 1px solid #666;
-  border-right: 0.5px solid #666;
-}
-
-.screen-window span {
-  margin-right: 10px;
-}
-
-.zero {
-  grid-column: 1/3;
-}
-
-.color-i {
-  background: rgb(246, 157, 2);
-  border: none;
-  border-bottom: 0.5px solid #666;
-}
